Move files about parser test
