+++ src/test/java/seedu/address/logic/parser/person/SelectCommandParserTest.java
@@ -1,4 +1,4 @@
-package seedu.address.logic.parser;
+package seedu.address.logic.parser.person;
 
 import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CommandParserTestUtil.assertParseFailure;
@@ -8,6 +8,7 @@ import static seedu.address.testutil.TypicalIndexes.INDEX_FIRST_PERSON;
 import org.junit.Test;
 
 import seedu.address.logic.commands.person.SelectCommand;
+import seedu.address.logic.parser.person.DeleteCommandParserTest;
 import seedu.address.logic.parser.person.SelectCommandParser;
 
 /**
